bug fix: check if file exists before removing
git-svn-id: http://svn.vdombox.ru/flextools/trunk@6880 230e8929-42e3-3b44-a413-55659dd3a577

+++ Help/Reader/v.01/src/Update.as
@@ -431,7 +431,9 @@ package
 			trace("Delete: " + location);
 			var newFileName:String = location;
 			var newFile:File = File.applicationStorageDirectory.resolvePath(newFileName);
-			newFile.deleteFile();
+			if (newFile.exists) {
+				newFile.deleteFile();
+			}
 //			trace(newFile.toString())
 		}
